Restructure PitchesController tests
