+++ rueppellii-lvlup-asp.net-core.IntegrationTests/rueppellii-lvlup-asp.net-core.IntegrationTests.csproj
@@ -23,8 +23,4 @@
     <ProjectReference Include="..\rueppellii-lvlup-asp.net-core\rueppellii-lvlup-asp.net-core.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Scenarios\PitchController\" />
-  </ItemGroup>
-
 </Project>
